save the prerequisite file on clicking execute

+++ SendLog.py
@@ -29,7 +29,7 @@ from xml.dom import minidom
 
 from db_connection import connection
 import config
-from config import FLASKSESSIONDIR, SESSIONDIR, XCOSSOURCEDIR
+from config import FLASKSESSIONDIR, SESSIONDIR, XCOSSOURCEDIR, REMOVEFILE
 
 
 def makedirs(dirname, dirtype):
@@ -41,6 +41,9 @@ def makedirs(dirname, dirtype):
 def remove(filename):
     if filename is None:
         return False
+    if not REMOVEFILE:
+        print("not removing", filename)
+        return True
     try:
         os.remove(filename)
         return True
@@ -97,7 +100,7 @@ DISPLAY_LIMIT = 10
 SCILAB_START = (
     "errcatch(-1,'stop');lines(0,120);clearfun('messagebox');"
     "function messagebox(msg,msgboxTitle,msgboxIcon,buttons,isModal),"
-    "disp(msg),endfunction;loadXcosLibs();")
+    "disp(msg),endfunction;")
 SCILAB_END = "mode(2);quit();"
 SCILAB_VARS = [
     "%p_r_p",
@@ -105,6 +108,7 @@ SCILAB_VARS = [
     "close",
     "extractDatatip",
     "extractLight",
+    "messagebox",
     "syslin",
     "tf2ss",
 ]
@@ -195,10 +199,6 @@ class line_and_state:
     '''
     line = None  # initial line to none(Nothing is present)
     state = NOLINE  # initial state to NOLINE ie
-
-    def __init__(self, line, state):
-        self.line = line
-        self.state = state
 
     def set(self, line_state):
         self.line = line_state[0]  # to set line
@@ -303,7 +303,7 @@ def add_script():
     return (script, sessiondir)
 
 
-def parse_line(line):
+def parse_line(line, lineno):
     '''
     Function to parse the line
     Returns tuple of figure ID and state
@@ -312,31 +312,35 @@ def parse_line(line):
             DATA otherwise
     '''
     line_words = line.split(' ')  # Each line is split to read condition
-    # The below condition determines the block ID
-    if line_words[2] == "Block":
-        # to get block id (Which is explicitly added by us while writing into
-        # log in scilab source code)
-        block_id = int(line_words[4])
-        return (block_id, BLOCK_IDENTIFICATION)
-    if line_words[2] == "Initialization":
-        # New figure created
-        # Get fig id
-        # to extract figure ids (sometime multiple sinks can be used in one
-        # diagram to differentiate that)
-        figure_id = int(line_words[-1])
-        return (figure_id, INITIALIZATION)
-    elif line_words[2] == "Ending":
-        # Current figure end
-        # Get fig id
-        figure_id = int(line_words[-1])
-        return (figure_id, ENDING)
-    else:
-        # Current figure coordinates
-        figure_id = int(line_words[3])
-        return (figure_id, DATA)
+    try:
+        # The below condition determines the block ID
+        if line_words[2] == "Block":
+            # to get block id (Which is explicitly added by us while writing
+            # into log in scilab source code)
+            block_id = int(line_words[4])
+            return (block_id, BLOCK_IDENTIFICATION)
+        if line_words[2] == "Initialization":
+            # New figure created
+            # Get fig id
+            # to extract figure ids (sometime multiple sinks can be used in one
+            # diagram to differentiate that)
+            figure_id = int(line_words[-1])
+            return (figure_id, INITIALIZATION)
+        elif line_words[2] == "Ending":
+            # Current figure end
+            # Get fig id
+            figure_id = int(line_words[-1])
+            return (figure_id, ENDING)
+        else:
+            # Current figure coordinates
+            figure_id = int(line_words[3])
+            return (figure_id, DATA)
+    except Exception as e:
+        print(str(e), "while parsing", line, "on line", lineno)
+        return (None, NOLINE)
 
 
-def get_line_and_state(file, figure_list):
+def get_line_and_state(file, figure_list, lineno):
     '''
     Function to get a new line from file
     This also parses the line and appends new figures to figure List
@@ -345,7 +349,7 @@ def get_line_and_state(file, figure_list):
     if not line:            # if line is empty then return noline
         return (None, NOLINE)
     # every line is passed to function parse_line for getting values
-    parse_result = parse_line(line)
+    parse_result = parse_line(line, lineno)
     figure_id = parse_result[0]
     state = parse_result[1]
     if state == INITIALIZATION:
@@ -363,6 +367,8 @@ def get_line_and_state(file, figure_list):
         # will be removed
         figure_list.remove(figure_id)
         return (None, ENDING)
+    elif state == NOLINE:
+        return (None, NOLINE)
     return (line, DATA)
 
 
@@ -734,6 +740,7 @@ def start_scilab():
     #    rather than Sinks's log file.
     #    0/No-condition : For all other blocks
 
+    command += "loadXcosLibs();"
     command += "importXcosDiagram('" + diagram.xcos_file_name + "');"
     command += "xcos_simulate(scs_m,4);"
 
@@ -813,6 +820,10 @@ def event_stream():
     This function is called in app route 'SendLog' below
     '''
     (diagram, __) = get_diagram(get_request_id())
+    if diagram is None:
+        print('no diagram')
+        yield "event: ERROR\ndata: no diagram\n\n"
+        return
 
     # Open the log file
     if not isfile(diagram.log_name):
@@ -830,9 +841,13 @@ def event_stream():
 
     with open(diagram.log_name, "r") as log_file:
         # Start sending log
-        line = line_and_state(None, NOLINE)
-        while line.set(get_line_and_state(log_file, diagram.figure_list)) or \
-                len(diagram.figure_list) > 0:
+        lineno = 0
+        line = line_and_state()
+        while True:
+            lineno += 1
+            line.set(get_line_and_state(log_file, diagram.figure_list, lineno))
+            if len(diagram.figure_list) == 0:
+                break
             # Get the line and loop until the state is ENDING and figure_list
             # empty. Determine if we get block id and give it to chart.js
             if line.get_state() == BLOCK_IDENTIFICATION:
@@ -841,8 +856,6 @@ def event_stream():
                 gevent.sleep(LOOK_DELAY)
             else:
                 yield "event: log\ndata: " + line.get_line() + "\n\n"
-            # Reset line, so server won't send same line twice
-            line = line_and_state(None, NOLINE)
 
     # Finished Sending Log
     kill_scilab(diagram)
